Refactor livepush code
